requirements.txt
Flask
psycopg2-binary
python-dotenv
schedule
tabulate